+++ NEWS.md
@@ -1,5 +1,6 @@
 ## propr 2.1.3.9000
 ---------------------
+* Add the `migraph` module to help make `igraph` networks
 * Penalize `lr2cor` and `aldex.cor` p-values for two-tailed test
 * Add `progress` bar to `aldex2propr` and `aldex.cor`